* Fixed a typo that prevented arenas from changing rating and giving wins * Hackfix: Shadowfury and Flamestrike don't drain all mana in client * Ported some melee hit chance fixes from AC * Trinkets Proc Fixes by Morpheux http://www.aspiredev.org/forums/viewtopic.php?f=22&t=24472 * Thunderfury doesn't debuff self patch by snail
git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/Hearthstone@1195 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/Player.cpp
@@ -7697,7 +7697,7 @@ void Player::DuelCountdown()
 
 	m_duelCountdownTimer -= 1000;
 
-	if( m_duelCountdownTimer < 0 )
+	if( int32(m_duelCountdownTimer) < 0 )
 		m_duelCountdownTimer = 0;
 
 	if( m_duelCountdownTimer == 0 )
